Separate the chain-index indexers into multiple individual parts
* Chain index memory and sqlite indexers are separated

* Refactored indexer modules for chain-index
