+++ plutus-chain-index-core/src/Plutus/ChainIndex/Emulator.hs
@@ -5,7 +5,7 @@ module Plutus.ChainIndex.Emulator(
 import Plutus.ChainIndex.ChainIndexError as Export
 import Plutus.ChainIndex.ChainIndexLog as Export
 import Plutus.ChainIndex.Effects as Export
-import Plutus.ChainIndex.Emulator.DiskState as Export hiding (fromTx)
-import Plutus.ChainIndex.Emulator.Handlers as Export
-import Plutus.ChainIndex.Emulator.Server as Export
+import Plutus.ChainIndex.Http.MemoryBackend.Server as Export
+import Plutus.ChainIndex.Indexer.Memory.DiskState as Export hiding (fromTx)
+import Plutus.ChainIndex.Indexer.Memory.Handlers as Export
 import Plutus.ChainIndex.Types as Export
