+++ plutus-chain-index-core/test/Plutus/ChainIndex/Indexer/Memory/HandlersSpec.hs
@@ -1,18 +1,17 @@
 {-# LANGUAGE DataKinds         #-}
 {-# LANGUAGE FlexibleContexts  #-}
 {-# LANGUAGE MonoLocalBinds    #-}
-{-# LANGUAGE NamedFieldPuns    #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE TupleSections     #-}
 {-# LANGUAGE TypeApplications  #-}
 
-module Plutus.ChainIndex.Emulator.HandlersSpec (tests) where
+module Plutus.ChainIndex.Indexer.Memory.HandlersSpec (tests) where
 
 import Control.Lens
 import Control.Monad (forM)
 import Control.Monad.Freer (Eff, interpret, reinterpret, runM)
 import Control.Monad.Freer.Error (Error, runError)
-import Control.Monad.Freer.Extras (LogMessage, LogMsg (..), handleLogWriter)
+import Control.Monad.Freer.Extras (LogMessage, LogMsg, handleLogWriter)
 import Control.Monad.Freer.State (State, runState)
 import Control.Monad.Freer.Writer (runWriter)
 import Control.Monad.IO.Class (liftIO)
@@ -24,12 +23,12 @@ import Generators qualified as Gen
 import Ledger (outValue)
 import Plutus.ChainIndex (ChainIndexLog, Page (pageItems), PageQuery (PageQuery), appendBlocks, unspentTxOutFromRef,
                           utxoSetWithCurrency)
-import Plutus.ChainIndex.Api (UtxosResponse (UtxosResponse))
 import Plutus.ChainIndex.ChainIndexError (ChainIndexError)
 import Plutus.ChainIndex.Effects (ChainIndexControlEffect, ChainIndexQueryEffect)
-import Plutus.ChainIndex.Emulator.Handlers (ChainIndexEmulatorState, handleControl, handleQuery)
+import Plutus.ChainIndex.Emulator (ChainIndexEmulatorState, handleChainIndexControl, handleQuery)
+import Plutus.ChainIndex.Http.Api (UtxosResponse (UtxosResponse))
 import Plutus.ChainIndex.Tx (_ValidTx, citxOutputs)
-import Plutus.ChainIndex.Types (ChainSyncBlock (..))
+import Plutus.ChainIndex.Types (ChainSyncBlock (Block))
 import Plutus.V1.Ledger.Value (AssetClass (AssetClass), flattenValue)
 
 import Hedgehog (Property, assert, forAll, property, (===))
@@ -61,7 +60,7 @@ eachTxOutRefAtAddressShouldBeUnspentSpec = property $ do
 
   result <- liftIO $ runEmulatedChainIndex mempty $ do
     -- Append the generated block in the chain index
-    appendBlocks [(Block tip (map (, def) block))]
+    appendBlocks [Block tip (map (, def) block)]
     utxoSetFromBlockAddrs block
 
   case result of
@@ -77,13 +76,13 @@ eachTxOutRefAtAddressShouldHaveTxOutSpec = property $ do
 
   result <- liftIO $ runEmulatedChainIndex mempty $ do
     -- Append the generated block in the chain index
-    appendBlocks [(Block tip (map (, def) block))]
+    appendBlocks [Block tip (map (, def) block)]
     utxos <- utxoSetFromBlockAddrs block
     traverse unspentTxOutFromRef (concat utxos)
 
   case result of
     Left _        -> Hedgehog.assert False
-    Right utxouts -> Hedgehog.assert $ and $ map isJust utxouts
+    Right utxouts -> Hedgehog.assert $ all isJust utxouts
 
 -- | After generating and appending a block in the chain index, verify that
 -- querying the chain index with each of the asset classes in the block returns
@@ -99,7 +98,7 @@ eachTxOutRefWithCurrencyShouldBeUnspentSpec = property $ do
 
   result <- liftIO $ runEmulatedChainIndex mempty $ do
     -- Append the generated block in the chain index
-    appendBlocks [(Block tip (map (, def) block))]
+    appendBlocks [Block tip (map (, def) block)]
 
     forM assetClasses $ \ac -> do
       let pq = PageQuery 200 Nothing
@@ -119,7 +118,7 @@ cantRequestForTxOutRefsWithAdaSpec = property $ do
 
   result <- liftIO $ runEmulatedChainIndex mempty $ do
     -- Append the generated block in the chain index
-    appendBlocks [(Block tip (map (, def) block))]
+    appendBlocks [Block tip (map (, def) block)]
 
     let pq = PageQuery 200 Nothing
     UtxosResponse _ utxoRefs <- utxoSetWithCurrency pq (AssetClass ("", ""))
@@ -142,7 +141,7 @@ runEmulatedChainIndex
   -> IO (Either ChainIndexError a)
 runEmulatedChainIndex appState effect = do
   r <- effect
-    & interpret handleControl
+    & handleChainIndexControl
     & interpret handleQuery
     & runState appState
     & runError
